+++ source/Playnite.DesktopApp/Controls/Views/SearchWindow.cs
@@ -120,7 +120,7 @@ namespace Playnite.DesktopApp.Controls.Views
                 BindingTools.SetBinding(
                     TextName,
                     TextBlock.TextProperty,
-                    GetBindingPath(nameof(GamesCollectionViewEntry.Name)),
+                    GetBindingPath(nameof(GamesCollectionViewEntry.DisplayName)),
                     BindingMode.OneTime);
             }
 
@@ -178,8 +178,7 @@ namespace Playnite.DesktopApp.Controls.Views
                 BindingTools.SetBinding(
                     ContentIcon,
                     ContentControl.ContentProperty,
-                    nameof(SearchItemWrapper.ItemIcon),
-                    BindingMode.OneTime);
+                    nameof(SearchItemWrapper.ItemIcon));
             }
 
             TextName = Template.FindName("PART_TextName", this) as TextBlock;