+++ source/Playnite/Settings/PlayniteSettings.cs
@@ -2136,6 +2136,15 @@ namespace Playnite
         private bool updateNotificationOnPatchesOnly = false;
         public bool UpdateNotificationOnPatchesOnly { get => updateNotificationOnPatchesOnly; set => SetValue(ref updateNotificationOnPatchesOnly, value); }
 
+        private string webImageSarchIconTerm = "{Name} icon";
+        public string WebImageSarchIconTerm { get => webImageSarchIconTerm; set => SetValue(ref webImageSarchIconTerm, value); }
+
+        private string webImageSarchCoverTerm = "{Name} cover";
+        public string WebImageSarchCoverTerm { get => webImageSarchCoverTerm; set => SetValue(ref webImageSarchCoverTerm, value); }
+
+        private string webImageSarchBackgroundTerm = "{Name} wallpaper";
+        public string WebImageSarchBackgroundTerm { get => webImageSarchBackgroundTerm; set => SetValue(ref webImageSarchBackgroundTerm, value); }
+
         [JsonIgnore]
         public static bool IsPortable
         {