+++ source/PlayniteSDK/Models/Emulator.cs
@@ -10,6 +10,97 @@ using System.Threading.Tasks;
 namespace Playnite.SDK.Models
 {
     /// <summary>
+    ///
+    /// </summary>
+    public class EmulatorDefinitionProfile
+    {
+        /// <summary>
+        ///
+        /// </summary>
+        public string Name { get; set; }
+
+        /// <summary>
+        ///
+        /// </summary>
+        public List<string> Platforms { get; set; }
+
+        /// <summary>
+        ///
+        /// </summary>
+        public List<string> ImageExtensions { get; set; }
+
+        /// <summary>
+        ///
+        /// </summary>
+        public List<string> ProfileFiles { get; set; }
+
+        /// <summary>
+        ///
+        /// </summary>
+        public string InstallationFile { get; set; }
+
+        /// <summary>
+        ///
+        /// </summary>
+        public string StartupArguments { get; set; }
+
+        /// <summary>
+        ///
+        /// </summary>
+        public string StartupExecutable { get; set; }
+
+        /// <summary>
+        ///
+        /// </summary>
+        public bool ScriptStartup { get; set; }
+
+        /// <summary>
+        ///
+        /// </summary>
+        public bool ScriptGameImport { get; set; }
+
+        /// <inheritdoc/>
+        public override string ToString()
+        {
+            return Name;
+        }
+    }
+
+    /// <summary>
+    ///
+    /// </summary>
+    public class EmulatorDefinition
+    {
+        internal string DirectoryName { get; set; }
+
+        /// <summary>
+        ///
+        /// </summary>
+        public string Id { get; set; }
+
+        /// <summary>
+        ///
+        /// </summary>
+        public string Name { get; set; }
+
+        /// <summary>
+        ///
+        /// </summary>
+        public string Website { get; set; }
+
+        /// <summary>
+        ///
+        /// </summary>
+        public List<EmulatorDefinitionProfile> Profiles { get; set; }
+
+        /// <inheritdoc/>
+        public override string ToString()
+        {
+            return Name;
+        }
+    }
+
+    /// <summary>
     /// Represents built-in region definition.
     /// </summary>
     public class EmulatedRegion : IEquatable<EmulatedRegion>